requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
python-dotenv==1.0.0
openai>=1.1.1
requests==2.31.0
reportlab==4.0.0
gunicorn==21.2.0